move TestMain
